Fix instance and lifecycle tests
