+++ web/apps/apps_test.go
@@ -19,6 +19,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/cozy/cozy-stack/pkg/instance/lifecycle"
 	"github.com/cozy/cozy-stack/pkg/oauth"
 	"github.com/cozy/cozy-stack/pkg/statik/fs"
 
@@ -306,7 +307,7 @@ func TestFaviconWithContext(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Test the theme
-	instance.Patch(testInstance, &instance.Options{
+	lifecycle.Patch(testInstance, &lifecycle.Options{
 		ContextName: context,
 	})
 	assert.NoError(t, err)
@@ -546,7 +547,7 @@ func TestMain(m *testing.M) {
 	cfg.Subdomains = config.NestedSubdomains
 	defer func() { cfg.Subdomains = was }()
 
-	testInstance = setup.GetTestInstance(&instance.Options{Domain: domain})
+	testInstance = setup.GetTestInstance(&lifecycle.Options{Domain: domain})
 	pass := "aephe2Ei"
 	hash, _ := crypto.GenerateFromPassphrase([]byte(pass))
 	testInstance.PassphraseHash = hash
